+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autos/fCamBlueFarAuto.java
@@ -1,8 +1,9 @@
-package org.firstinspires.ftc.teamcode.robot;
+package org.firstinspires.ftc.teamcode.autos;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.geometry.Vector2d;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
@@ -11,29 +12,32 @@ import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
 import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
 import org.firstinspires.ftc.teamcode.pipelines.ColorMask;
+import org.firstinspires.ftc.teamcode.robot.HwMap;
+import org.firstinspires.ftc.teamcode.robot.RobotConstants;
 import org.firstinspires.ftc.teamcode.trajectorysequence.TrajectorySequence;
 import org.openftc.easyopencv.OpenCvCamera;
 import org.openftc.easyopencv.OpenCvCameraFactory;
 import org.openftc.easyopencv.OpenCvCameraRotation;
 
-@Autonomous(name = "fCam Blue Close Auto")
-public class fCamBlueCloseAuto extends LinearOpMode {
+@Disabled
+@Autonomous(name = "fCam Blue Far Auto")
+public class fCamBlueFarAuto extends LinearOpMode {
 
     enum Camera {
         WAIT,
         SAVE,
         FINISHED
     }
-    fCamBlueFarAuto.Camera camera = fCamBlueFarAuto.Camera.WAIT;
+    Camera camera = Camera.WAIT;
 
 //    enum AutoPath {
 //        LEFT,
 //        CENTER,
 //        RIGHT
 //    }
-//    AutoPath autoPath = AutoPath.RIGHT;
+//    AutoPath autoPath = AutoPath.LEFT;
 
-    int targetSlidePos = RobotConstants.slideAuto;
+    int targetSlidePos = RobotConstants.slideLow;
 
     double slideI = 0;
 
@@ -41,7 +45,6 @@ public class fCamBlueCloseAuto extends LinearOpMode {
 
     @Override
     public void runOpMode() throws InterruptedException {
-
 
         SampleMecanumDrive driveTrain = new SampleMecanumDrive(hardwareMap);
         ColorMask pipeline = new ColorMask();
@@ -67,58 +70,74 @@ public class fCamBlueCloseAuto extends LinearOpMode {
             }
         });
 
-        TrajectorySequence left = driveTrain.trajectorySequenceBuilder(new Pose2d(12, 63, Math.toRadians(270)))
-                .lineTo(new Vector2d(12, 61))
-                .splineToLinearHeading(new Pose2d(25, 45), Math.toRadians(270))
-                .addSpatialMarker(new Vector2d(25, 45), () -> robot.rightServo.setPosition(RobotConstants.rightIn))
+        TrajectorySequence left = driveTrain.trajectorySequenceBuilder(new Pose2d(-35, 63, Math.toRadians(270)))
+                .addDisplacementMarker(() -> targetSlidePos = RobotConstants.slideBottom)
+                .lineTo(new Vector2d(-35, 60))
+                .splineToSplineHeading(new Pose2d(-30, 37), Math.toRadians(0))
+                .splineToConstantHeading(new Vector2d(-22.5, 40.5), Math.toRadians(0))
+                .addSpatialMarker(new Vector2d(-22.5, 38), () -> robot.rightServo.setPosition(RobotConstants.rightIn))
                 .waitSeconds(.2)
-                .lineTo(new Vector2d(25, 46))
-                .lineTo(new Vector2d(26, 46))
-                .splineToConstantHeading(new Vector2d(45, 44), Math.toRadians(0))
-                .lineTo(new Vector2d(54.5, 44))
-                .lineTo(new Vector2d(54.6, 44))
-                .addSpatialMarker(new Vector2d(54.6, 44), () -> robot.dropServo.setPosition(RobotConstants.dropOpen))
-                .waitSeconds(.2)
-                .lineTo(new Vector2d(40, 44))
+                .splineToConstantHeading(new Vector2d(-12, 40), Math.toRadians(0))
+                //wait so that other team can run their auto
+                .waitSeconds(10)
+                .lineTo(new Vector2d(30, 40))
+                .splineToConstantHeading(new Vector2d(45, 42), Math.toRadians(0))
+                .addSpatialMarker(new Vector2d(35, 39), () -> targetSlidePos = RobotConstants.slideLow)
+                .lineTo(new Vector2d(54.5, 42))
+                .lineTo(new Vector2d(54.6, 42))
+                .addSpatialMarker(new Vector2d(54.6, 42), () -> robot.dropServo.setPosition(RobotConstants.dropOpen))
+                .waitSeconds(.3)
+                .lineTo(new Vector2d(40, 42))
                 .addDisplacementMarker(() -> {targetSlidePos = RobotConstants.slideBottom; robot.dropServo.setPosition(RobotConstants.dropClosed);})
-                .lineTo(new Vector2d(40, 60))
-                .lineTo(new Vector2d(45, 60))
+                .lineTo(new Vector2d(40, 10))
+                .lineTo(new Vector2d(45, 10))
                 .build();
 
-        TrajectorySequence center = driveTrain.trajectorySequenceBuilder(new Pose2d(12, 63, Math.toRadians(270)))
-                .lineTo(new Vector2d(12, 60))
-                .splineToSplineHeading(new Pose2d(17, 30), Math.toRadians(270))
-                .lineTo(new Vector2d(17,36))
-                .addTemporalMarker(2.2, () -> robot.rightServo.setPosition(RobotConstants.rightIn))
-//                                .addSpatialMarker(new Vector2d(15, 35), () -> robot.rightServo.setPosition(RobotConstants.rightIn))
+        TrajectorySequence center = driveTrain.trajectorySequenceBuilder(new Pose2d(-35, 63, Math.toRadians(270)))
+                .addDisplacementMarker(() -> targetSlidePos = RobotConstants.slideBottom)
+                .lineTo(new Vector2d(-35, 60))
+                .splineToSplineHeading(new Pose2d(-33, 35), Math.toRadians(285))
+                .addSpatialMarker(new Vector2d(-33, 35), () -> robot.rightServo.setPosition(RobotConstants.rightIn))
                 .waitSeconds(.2)
-                .lineTo(new Vector2d(18, 36))
-                .splineToConstantHeading(new Vector2d(45, 37.5), Math.toRadians(0))
-                .lineTo(new Vector2d(54.5, 37.5))
-                .lineTo(new Vector2d(54.6, 37.5))
-                .addSpatialMarker(new Vector2d(54.6, 37.5), () -> robot.dropServo.setPosition(RobotConstants.dropOpen))
+                .splineToConstantHeading(new Vector2d(-27, 36), Math.toRadians(0))
+                .lineTo(new Vector2d(-12, 36))
+                //wait so that other team can run their auto
+                .waitSeconds(10)
+                .lineTo(new Vector2d(30, 36))
+                .splineToConstantHeading(new Vector2d(45, 35), Math.toRadians(0))
+                .addSpatialMarker(new Vector2d(38, 35), () -> targetSlidePos = RobotConstants.slideLow)
+                .lineTo(new Vector2d(54.5, 35))
+                .lineTo(new Vector2d(54.6, 35))
+                .addSpatialMarker(new Vector2d(54.6, 35), () -> robot.dropServo.setPosition(RobotConstants.dropOpen))
                 .waitSeconds(.3)
-                .lineTo(new Vector2d(40, 37.5))
+                .lineTo(new Vector2d(40, 35))
                 .addDisplacementMarker(() -> {targetSlidePos = RobotConstants.slideBottom; robot.dropServo.setPosition(RobotConstants.dropClosed);})
-                .lineTo(new Vector2d(40, 60))
-                .lineTo(new Vector2d(45, 60))
+                .lineTo(new Vector2d(40, 10))
+                .lineTo(new Vector2d(45, 10))
                 .build();
 
-        TrajectorySequence right = driveTrain.trajectorySequenceBuilder(new Pose2d(12, 63, Math.toRadians(270)))
-                .lineTo(new Vector2d(12, 45))
-                .splineToConstantHeading(new Vector2d(8.5, 36), Math.toRadians(270))
-                .addSpatialMarker(new Vector2d(8.5, 36), () -> robot.rightServo.setPosition(RobotConstants.rightIn))
+        TrajectorySequence right = driveTrain.trajectorySequenceBuilder(new Pose2d(-35, 63, Math.toRadians(270)))
+                .addDisplacementMarker(() -> targetSlidePos = RobotConstants.slideBottom)
+                .lineTo(new Vector2d(-35, 60))
+                .splineToConstantHeading(new Vector2d(-38, 34), Math.toRadians(270))
+                .addSpatialMarker(new Vector2d(-38,34), () -> robot.rightServo.setPosition(RobotConstants.rightIn))
                 .waitSeconds(.2)
-                .lineTo(new Vector2d(20, 36))
-                .splineToSplineHeading(new Pose2d(45, 31, Math.toRadians(0)), Math.toRadians(0))
-                .lineTo(new Vector2d(54.5, 29.5))
-                .lineTo(new Vector2d(54.6, 29.5))
-                .addSpatialMarker(new Vector2d(54.6, 29.5), () -> robot.dropServo.setPosition(RobotConstants.dropOpen))
+                .lineTo(new Vector2d(-33, 34))
+                .splineToSplineHeading(new Pose2d(-28, 35, Math.toRadians(0)), Math.toRadians(0))
+                .lineTo(new Vector2d(-12, 35))
+                //wait so that other team can run their auto
+                .waitSeconds(10)
+                .lineTo(new Vector2d(30, 35))
+                .splineToConstantHeading(new Vector2d(45, 23), Math.toRadians(0))
+                .addSpatialMarker(new Vector2d(38, 36), () -> targetSlidePos = RobotConstants.slideLow)
+                .lineTo(new Vector2d(54.5, 23))
+                .lineTo(new Vector2d(54.6, 23))
+                .addSpatialMarker(new Vector2d(54.6, 23), () -> robot.dropServo.setPosition(RobotConstants.dropOpen))
                 .waitSeconds(.3)
-                .lineTo(new Vector2d(40, 29.5))
+                .lineTo(new Vector2d(40, 23))
                 .addDisplacementMarker(() -> {targetSlidePos = RobotConstants.slideBottom; robot.dropServo.setPosition(RobotConstants.dropClosed);})
-                .lineTo(new Vector2d(40, 60))
-                .lineTo(new Vector2d(45, 60))
+                .lineTo(new Vector2d(40, 10))
+                .lineTo(new Vector2d(45, 10))
                 .build();
 
         ElapsedTime cameraDelayTimer = new ElapsedTime();
@@ -128,6 +147,8 @@ public class fCamBlueCloseAuto extends LinearOpMode {
 
         Telemetry.Item slideData = telemetry.addData("Slide Data:", "Encoder Val:" + robot.liftEncoder.getCurrentPosition() + " Target Val:" + targetSlidePos);
 
+
+
         robot.rightServo.setPosition(RobotConstants.rightOut);
         robot.dropServo.setPosition(RobotConstants.dropClosed);
 
@@ -136,7 +157,7 @@ public class fCamBlueCloseAuto extends LinearOpMode {
 
         robot.climbMotor.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
 
-        driveTrain.setPoseEstimate(new Pose2d(12, 63, Math.toRadians(270)));
+        driveTrain.setPoseEstimate(new Pose2d(-35, 63, Math.toRadians(270)));
 
         cameraDelayTimer.reset();
 
@@ -146,7 +167,7 @@ public class fCamBlueCloseAuto extends LinearOpMode {
             switch (camera) {
                 case WAIT:
                     if (cameraDelayTimer.seconds() > 1.5) {
-                        camera = fCamBlueFarAuto.Camera.SAVE;
+                        camera = Camera.SAVE;
                     }
                     break;
                 case SAVE:
@@ -169,7 +190,7 @@ public class fCamBlueCloseAuto extends LinearOpMode {
                             detectedPos.setValue("Default center (No detection)");
                             break;
                     }
-                    camera = fCamBlueFarAuto.Camera.FINISHED;
+                    camera = Camera.FINISHED;
                     break;
                 case FINISHED:
                     break;
